Fix About Page Image Stretching on Mobile
Changed style="height:500px" to style="max-height:500px". This fixes an issue where on small screens, the image height was forced to be 500px, but to scale down and fit the page width, it should have been less than 500px. Setting "max-height" instead of "height" allows the image to scale down further and be less than 500px tall if need be.